move relevant functions to a new utils_3d.py
